Fix cmd to enable key-repeating in Vim mode on Mac
The original command in the markdown read:
`defaults write com.insomnia.app ApplePressAndHoldEnabled -bool false<br>` and I couldn't run it from zsh. I got this output: 
`zsh: parse error near \n'
when I got rid of the `<br>`, it worked. I restarted Insomnia and had the key-repeating I wanted, so I regarded this as a typo, probably due to page layout needs. Just in case it's for the layout, I added a newline below the command, outside the backticks.